Align with other api. fix perf regression where groups were re-evaluated multi times in completeness queries
